Check if an object is still in index_queryset
There is an issue where if an object is updated in a way that removes it from the index_queryset it does not get removed upon update.  This can be fixed by overriding update_object or instead should_update to check if an object is no longer in the queryset and removing instead of updating.